Swap over to use tracing
