+++ plugin-server/src/common/tracing/otel.ts
@@ -1,4 +1,3 @@
-import { Attributes, SpanKind, SpanStatusCode, Tracer, trace } from '@opentelemetry/api'
 import { getNodeAutoInstrumentations } from '@opentelemetry/auto-instrumentations-node'
 import { OTLPTraceExporter } from '@opentelemetry/exporter-trace-otlp-grpc'
 import { PgInstrumentation } from '@opentelemetry/instrumentation-pg'
@@ -9,8 +8,6 @@ import { ATTR_SERVICE_NAME, ATTR_SERVICE_VERSION } from '@opentelemetry/semantic
 
 import { defaultConfig } from '~/config/config'
 import { logger } from '~/utils/logger'
-
-export const tracer = trace.getTracer('postgres')
 
 function redactSQL(sql: string): string {
     return sql
@@ -66,26 +63,4 @@ export const initTracing = (): void => {
         })
         sdk.start()
     }
-}
-
-export function withSpan<T>(
-    tracer: Tracer | string,
-    name: string,
-    attrs: Attributes,
-    fn: () => Promise<T>
-): Promise<T> {
-    const _tracer = typeof tracer === 'string' ? trace.getTracer(tracer) : tracer
-    return _tracer.startActiveSpan(name, { kind: SpanKind.CLIENT, attributes: attrs }, async (span) => {
-        try {
-            const out = await fn()
-            span.setStatus({ code: SpanStatusCode.OK })
-            return out
-        } catch (e: any) {
-            span.recordException(e)
-            span.setStatus({ code: SpanStatusCode.ERROR, message: e?.message })
-            throw e
-        } finally {
-            span.end()
-        }
-    })
 }
