+++ plugin-server/src/index.ts
@@ -1,5 +1,5 @@
 // NOTE: Keep these as ~ imports as we can validate the build output this way
-import { initTracing } from '~/common/otel'
+import { initTracing } from '~/common/tracing/otel'
 import { PluginServer } from '~/server'
 import { initSuperProperties } from '~/utils/posthog'
