+++ plugin-server/tests/helpers/clickhouse.ts
@@ -2,6 +2,7 @@ import { ClickHouseClient, ExecResult, createClient as createClickhouseClient } 
 import { performance } from 'perf_hooks'
 import { Readable } from 'stream'
 
+import { withSpan } from '~/common/tracing/tracing-utils'
 import {
     ClickHouseEvent,
     ClickHousePerson,
@@ -16,7 +17,6 @@ import { timeoutGuard } from '~/utils/db/utils'
 import { isTestEnv } from '~/utils/env-utils'
 import { parseRawClickHouseEvent } from '~/utils/event'
 import { parseJSON } from '~/utils/json-parse'
-import { instrumentQuery } from '~/utils/metrics'
 import { fetch } from '~/utils/request'
 
 import { logger } from '../../src/utils/logger'
@@ -160,7 +160,7 @@ export class Clickhouse {
     }
 
     query<T>(query: string): Promise<T[]> {
-        return instrumentQuery('query.clickhouse', undefined, async () => {
+        return withSpan('clickhouse', 'query.clickhouse', { tag: 'unknown' }, async () => {
             const timeout = timeoutGuard('ClickHouse slow query warning after 30 sec', { query })
             try {
                 const queryResult = await this.client.query({
